Refactor report creation
