+++ app/reports/index.js
@@ -1,4 +1,4 @@
-const { createMIReport } = require('./mi-report')
+const { createMIReport } = require('./mi-report/create')
 
 module.exports = {
   createMIReport
