+++ app/reports/mi-report/convert-to-csv.js
@@ -1,0 +1,10 @@
+const convertToCSV = (data) => {
+  let csv = ''
+  csv = data.map(row => Object.values(row))
+  csv.unshift(Object.keys(data[0]))
+  return `"${csv.join('"\n"').replace(/,/g, '","')}"`
+}
+
+module.exports = {
+  convertToCSV
+}
