+++ app/reports/mi-report/get-batch-export-date.js
@@ -1,0 +1,13 @@
+const moment = require('moment')
+
+const getBatchExportDate = (events) => {
+  const extractedEvent = events.find(event => event.type === 'uk.gov.defra.ffc.pay.payment.extracted')
+  if (extractedEvent) {
+    return moment(extractedEvent.time).format('DD/MM/YYYY')
+  }
+  return ''
+}
+
+module.exports = {
+  getBatchExportDate
+}
