+++ app/reports/mi-report/index.js
@@ -1,0 +1,5 @@
+const { createMIReport } = require('./create')
+
+module.exports = {
+  createMIReport
+}
