+++ app/reports/mi-report/order-grouped-events.js
@@ -1,0 +1,17 @@
+const { getEventOrder } = require('./get-event-order')
+
+const orderGroupedEvents = (events) => {
+  return events.map(group => {
+    const sortedEvents = group.events.sort((a, b) => {
+      return getEventOrder(a.type) - getEventOrder(b.type)
+    })
+    return {
+      correlationId: group.correlationId,
+      events: sortedEvents
+    }
+  })
+}
+
+module.exports = {
+  orderGroupedEvents
+}
